Refactor ConcurrentOutput test
